feature: TRACEFOSS-1401 refactored investigation structure.
