+++ tx-backend/src/integration/groovy/org/eclipse/tractusx/traceability/common/support/AssetsSupport.groovy
@@ -21,10 +21,10 @@
 
 package org.eclipse.tractusx.traceability.common.support
 
-import net.bytebuddy.build.BuildLogger
+
 import org.eclipse.tractusx.traceability.assets.infrastructure.adapters.feign.irs.model.AssetsConverter
 import org.eclipse.tractusx.traceability.assets.infrastructure.adapters.jpa.asset.AssetEntity
-import org.eclipse.tractusx.traceability.infrastructure.jpa.investigation.InvestigationEntity
+import org.eclipse.tractusx.traceability.investigations.adapters.jpa.InvestigationEntity
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationSide
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationStatus
 import org.slf4j.Logger
