+++ tx-backend/src/integration/groovy/org/eclipse/tractusx/traceability/common/support/NotificationsRepositoryProvider.groovy
@@ -22,7 +22,7 @@
 package org.eclipse.tractusx.traceability.common.support
 
 
-import org.eclipse.tractusx.traceability.infrastructure.jpa.notification.JpaNotificationRepository
+import org.eclipse.tractusx.traceability.investigations.adapters.jpa.JpaNotificationRepository
 
 interface NotificationsRepositoryProvider {
 	JpaNotificationRepository jpaNotificationRepository()
