+++ tx-backend/src/integration/groovy/org/eclipse/tractusx/traceability/infrastructure/edc/blackbox/EdcControllerIT.groovy
@@ -28,8 +28,8 @@ import org.eclipse.tractusx.traceability.common.support.InvestigationsSupport
 import org.eclipse.tractusx.traceability.common.support.NotificationsSupport
 import org.eclipse.tractusx.traceability.common.support.TestDataSupport
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.model.EDCNotification
-import org.eclipse.tractusx.traceability.infrastructure.jpa.investigation.InvestigationEntity
-import org.eclipse.tractusx.traceability.infrastructure.jpa.notification.NotificationEntity
+import org.eclipse.tractusx.traceability.investigations.adapters.jpa.InvestigationEntity
+import org.eclipse.tractusx.traceability.investigations.adapters.jpa.NotificationEntity
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationSide
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationStatus
 import org.eclipse.tractusx.traceability.investigations.domain.model.Severity
