+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/investigations/adapters/jpa/InvestigationEntity.java
@@ -19,10 +19,9 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.traceability.infrastructure.jpa.investigation;
+package org.eclipse.tractusx.traceability.investigations.adapters.jpa;
 
 import org.eclipse.tractusx.traceability.assets.infrastructure.adapters.jpa.asset.AssetEntity;
-import org.eclipse.tractusx.traceability.infrastructure.jpa.notification.NotificationEntity;
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationSide;
 import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationStatus;
